requirements.txt
fastapi   #фреймворк для створення нашого API (бекенду).#
uvicorn[standard]   # сервер, який буде запускати FastAPI.
sqlmodel  #зручна бібліотека для роботи з базою (ORM на основі SQLAlchemy + Pydantic).#
sqlalchemy[asyncio]  #“двигун” для роботи з базами, потрібен sqlmodel.#
aiosqlite  #“двигун” для роботи з базами, потрібен sqlmodel.#
pydantic #бібліотека для валідації та роботи зі схемами даних (FastAPI теж її використовує).#
aiogram==3.* # — бібліотека для створення Telegram-ботів.#
httpx #HTTP-клієнт (наш бот буде через нього “спілкуватися” з API).#
python-dotenv #щоб читати .env файл і підставляти змінні в код.#